Extract getReportAlternateText into shared module
